Add ability to return results of a module run for the RPC interface.
This is accomplished through a decorator and an internal list that
is part of the module object. Every time add_host() is called, the
data is copied to the result_cache list. When the module returns,
the decorator takes over and returns the result_cache and then
clears it for the next module run.